Updated date and time variables in scenes table to use string datatypes, and updated tracking variable in user_scenes table to an integer datatype.

+++ db/migrate/20140729212307_add_tracking_to_user_scenes.rb
@@ -1,5 +1,5 @@
 class AddTrackingToUserScenes < ActiveRecord::Migration
   def change
-    add_column :user_scenes, :tracking, :float
+    add_column :user_scenes, :tracking, :integer
   end
 end
